fixed bug in consumer

+++ src/Consumer.cpp
@@ -118,7 +118,7 @@ void ConsumerImpl::subscribe() {
                 auto& rpc = client->m_consumer_request_events;
                 auto& ph = partition.self->m_ph;
                 auto consumer_ptr = reinterpret_cast<intptr_t>(this);
-                Result<void> result = rpc.on(ph)(consumer_ptr, i, m_name, 0, 0);
+                Result<void> result = rpc.on(ph)(consumer_ptr, (size_t)i, m_name, (size_t)0, (size_t)0);
                 ult_completed[i].set_value();
             }
         );
